Get rid of all the useless resources
The library used to expose resources that are not part of the SDK itself (ic_launcher icons, colours, etc.). The SDK is mainly made of a Java class and hence could be exposed as a regular jar file. Let’s stick to the AAR format for future proofness but clean the exposed APIs/resources.